+++ libsnark/gadgetlib1/gadgets/merkle_tree/merkle_tree_check_read_gadget.tcc
@@ -14,6 +14,8 @@
 #ifndef MERKLE_TREE_CHECK_READ_GADGET_TCC_
 #define MERKLE_TREE_CHECK_READ_GADGET_TCC_
 
+#include <msvc_hack.h>
+
 namespace libsnark {
 
 template<typename FieldT, typename HashT>
@@ -99,7 +101,7 @@ template<typename FieldT, typename HashT>
 void merkle_tree_check_read_gadget<FieldT, HashT>::generate_r1cs_witness()
 {
     /* do the hash computations bottom-up */
-    for (int i = tree_depth-1; i >= 0; --i)
+    for (ssize_t i = tree_depth-1; i >= 0; --i)
     {
         /* propagate previous input */
         propagators[i].generate_r1cs_witness();
@@ -144,10 +146,10 @@ void test_merkle_tree_check_read_gadget()
     libff::bit_vector address_bits;
 
     size_t address = 0;
-    for (long level = tree_depth-1; level >= 0; --level)
+    for (ssize_t level = tree_depth-1; level >= 0; --level)
     {
         const bool computed_is_right = (std::rand() % 2);
-        address |= (computed_is_right ? 1ul << (tree_depth-1-level) : 0);
+        address |= (computed_is_right ? (size_t)1 << (tree_depth-1-level) : 0);
         address_bits.push_back(computed_is_right);
         libff::bit_vector other(digest_len);
         std::generate(other.begin(), other.end(), [&]() { return std::rand() % 2; });
